backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
python-multipart
pydantic
pydantic-settings
celery
redis
playwright
selenium
beautifulsoup4
pandas
numpy
python-dotenv
httpx
pytest
pytest-asyncio
